model verison upgrade to 19

+++ health-services/transformer/src/main/java/org/egov/transformer/Constants.java
@@ -1,7 +1,5 @@
 package org.egov.transformer;
 
-import org.apache.kafka.common.protocol.types.Field;
-
 public interface Constants {
     String PROJECT_TYPES = "projectTypes";
     String INDIVIDUAL = "INDIVIDUAL";
@@ -12,7 +10,7 @@ public interface Constants {
 
     String QUANTITY_WASTED = "quantityWasted";
 
-    Double RE_ADMINISTERED_DOSES = 2.0;
+    Long RE_ADMINISTERED_DOSES = 2l;
 
     String BENEFICIARY_REFERRED = "BENEFICIARY_REFERRED";
     String TASK_STATUS = "taskStatus";
